Selvana

+++ src/app/modules/fi/index/fi-account-hierarchy/fi-account-hierarchy.component.ts
@@ -6,7 +6,8 @@ import {MatPaginator, MatPaginatorModule} from '@angular/material/paginator';
 import {MatSort, MatSortModule} from '@angular/material/sort';
 import {MatTableDataSource, MatTableModule} from '@angular/material/table';
 import { FIAccountHierarchyDialogComponent } from '../fi-account-hierarchy-dialog/fi-account-hierarchy-dialog.component';
-
+import { HotkeysService } from 'angular2-hotkeys';
+import { Hotkey } from 'angular2-hotkeys';
 @Component({
   selector: 'app-fi-account-hierarchy',
   templateUrl: './fi-account-hierarchy.component.html',
@@ -19,9 +20,14 @@ export class FIAccountHierarchyComponent {
 
   @ViewChild(MatPaginator) paginator!: MatPaginator;
   @ViewChild(MatSort) sort!: MatSort;
-  constructor(private dialog : MatDialog, private api : ApiService){}
+  constructor(private dialog : MatDialog, private api : ApiService,private hotkeysService: HotkeysService){}
   ngOnInit(): void {
     this.getFIAccountHierarchies();
+    this.hotkeysService.add(new Hotkey('ctrl+o', (event: KeyboardEvent): boolean => {
+      // Call the deleteGrade() function in the current component
+      this.openDialog();
+      return false; // Prevent the default browser behavior
+    }));
   }
   openDialog() {
     this.dialog.open(FIAccountHierarchyDialogComponent, {
@@ -57,20 +63,28 @@ export class FIAccountHierarchyComponent {
     })
   }
   daleteFIAccountHierarchies(id:number){
-    if(confirm("Are you sure to delete ")) {
-      console.log("Implement delete functionality here");
+    var result = confirm('هل ترغب بتاكيد الحذف ؟ ');
+    if (result) {
+      this.api.deleteFIAccountHierarchy(id)
+      .subscribe({
+        next: (res) => {
+          if(res == 'Succeeded'){
+            console.log("res of deletestore:",res)
+          alert('تم الحذف بنجاح');
+          this.getFIAccountHierarchies();
+  
+        }else{
+          alert(" لا يمكن الحذف لارتباطها بجداول اخري!")
+        }
+        },
+        error: () => {
+          alert('خطأ فى حذف العنصر'); 
+        },
+      });
     }
-    this.api.deleteFIAccountHierarchy(id)
-    .subscribe({
-      next:(res)=>{
-        alert("تأكيد حذف الوحدة");
-        this.getFIAccountHierarchies();
-      },
-      error:()=>{
-        alert("خطأ عند الحذف")
-      }
-    })
   }
+
+  
 
   async getSearchGrades(level :any,name:any) {
